+++ CHANGES.md
@@ -9,7 +9,15 @@
 * #43 ESKF estimates vertical position and velocity
 * #43 Add IMU to the list of available sensors 
 * #43 Allow to define the frequency at which sensors correct/update ESKF estimates
+* #46 Implement handler for getting Motor values via MSP
+* #47 MSP getters for Mosquito version and presence of Position Board
+* #48 MSP message and methods to calibrate ESCs
+* #50 Check if position board is connected via MSP
 * #45 Allow to set rangefinder parameters with an MSP message
+* #55 Enable setting LEDs on/off with an MSP message
+* #68 Motor pins are set according to the specified Mosquito version
+* #70 Transmitter calibration
+* #73 Battery monitoring and low battery trigger 
 
 ### Changed
 
@@ -23,3 +31,6 @@
 
 * #39 Firmware version
 * #41 Serial communications with the platform
+* #49 Optical Flow connection pin
+* #51 Modify parameters at runtime (Mosquito version and Position Board presence)
+* #52 ESC calibration procedure